requirements.txt
python-dateutil
requests
numpy
netcdf4
pandas